+++ Dockerfile
@@ -11,7 +11,7 @@ COPY requirements.txt requirements.txt
 RUN pip install -r requirements.txt
 
 COPY . .
-
-CMD uvicorn main:app --reload --host 0.0.0.0 --port 8000
+EXPOSE 8000
+CMD uvicorn ocr_api:app --reload --host 0.0.0.0 --port 8000
 
 EXPOSE 8000
