+++ requirements.txt
@@ -53,7 +53,7 @@ tqdm==4.65.0
 triton==2.0.0
 typing_extensions==4.7.1
 tzdata==2023.3
-ultralytics==8.0.123
+ultralytics==8.0.149
 urllib3==1.26.16
 uvicorn==0.22.0
 Werkzeug==2.3.6
